Initial commit
This commit comes about 4 weeks into the project, but I wanted to get it
up here and update it as I continue working on it. Thus far, the
logistic regression function runs as expected. Some further parameter
tweaking and investigating may be necessary. TODO: Move on to 3
dimensional feature space training data.