Separated 2D data set and wrote 2D validator
I separated the 2D set into two portions based off of the line of best
fit, and ran the algorithm on each region separately. This allowed me to
create a 'band' of exoplanet candidacy instead of one dividing line. The
validator for this set is operational and returned a 60% correct
matching on the training set.